requirements.txt
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.0.4
django-allauth==0.51.0
djangorestframework==3.13.1
idna==3.4
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
Pillow==9.2.0
djangorestframework-simplejwt==5.2.1
MarkupSafe==2.1.1
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0
itypes-1.2.0 jinja2==3.1.2
openapi-codec==1.3.2
simplejson==3.17.6
uritemplate==4.1.1
drf-yasg==1.21.4
inflection==0.5.1
packaging==21.3
pyparsing==-3.0.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
attrs==22.1.0
jsonschema==4.16.0
pyrsistent==0.18.1
pyyaml==6.0
swagger-spec-validator==2.7.6
django-notifications-rest==0.0.1
notifications==0.3.2
cachecontrol==0.12.11
cachetools==5.2.0
fcm_django==1.0.12
firebase-admin==5.4.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
msgpack==1.0.4
proto-plus==1.22.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.9
django_crontab==0.7.1
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
kombu==5.2.4
prompt-toolkit==3.0.36
vine==5.0.0
wcwidth==0.2.5
async-timeout==4.0.2
django-celery-beat==2.4.0
django-celery-results==2.4.0
